Split performance tests from unit tests
